Move common / shared code into utils module and update affected scripts.
